# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Bike:
    properties:
      id:
        type: integer
      is_available:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      price_per_minute:
        type: number
    type: object
  models.Rental:
    properties:
      bike_id:
        type: integer
      cost:
        type: number
      created_at:
        type: string
      duration_minutes:
        type: integer
      end_latitude:
        type: number
      end_longitude:
        type: number
      end_time:
        type: string
      id:
        type: integer
      start_latitude:
        type: number
      start_longitude:
        type: number
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  types.EndRentalRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  types.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  types.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  types.StartRentalRequest:
    properties:
      bike_id:
        type: integer
    type: object
  types.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bikerental.com
    name: API Support
  description: API para gestión de alquiler de bicicletas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bike Rental Service API
  version: "1.0"
paths:
...
  /bikes/available:
    get:
      consumes:
      - application/json
      description: Get paginated list of available bikes for rent
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of available bikes
          schema:
            allOf:
            - $ref: '#/definitions/types.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Bike'
                  type: array
              type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List available bikes
      tags:
      - bikes
  /rentals/end:
    post:
      consumes:
      - application/json
      description: End the active rental for authenticated user with end location
      parameters:
      - description: End location coordinates
        in: body
        name: rental
        required: true
        schema:
          $ref: '#/definitions/types.EndRentalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rental ended successfully with cost
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Rental'
              type: object
        "400":
          description: Invalid coordinates or location too far from start
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "409":
          description: No active rental to end
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End a bike rental
      tags:
      - rentals
  /rentals/history:
    get:
      consumes:
      - application/json
      description: Get paginated rental history for authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rental history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Rental'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get rental history
      tags:
      - rentals
  /rentals/start:
    post:
      consumes:
      - application/json
      description: Start a new rental for the authenticated user with specified bike
      parameters:
      - description: Rental start data with bike_id
        in: body
        name: rental
        required: true
        schema:
          $ref: '#/definitions/types.StartRentalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rental started successfully
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Rental'
              type: object
        "400":
          description: Invalid request payload or bike_id
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Bike not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "409":
          description: User has active rental or bike not available
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a bike rental
      tags:
      - rentals
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"